backend/requirements.txt
# Flask and web framework dependencies
Flask==3.0.0
flask-cors==4.0.0

# HTTP requests
requests==2.31.0

# Environment variables
python-dotenv==1.0.0

# Development and testing
pytest==7.4.3
pytest-flask==1.3.0

# Production server (optional)
gunicorn==21.2.0

# Logging and monitoring
structlog==23.2.0